sql updated from int to numeric, allows decimals

+++ server/routes/hours.router.js
@@ -5,7 +5,7 @@ const router = express.Router();
 
 router.get('/weekly/:id', (req, res) => {
   const userId = req.user.id
-  queryText = `SELECT to_char(date_trunc('week', date ), 'MM/DD') AS "dates" , SUM (time_paid::INT)::VARCHAR "timePaid"
+  queryText = `SELECT to_char(date_trunc('week', date ), 'MM/DD') AS "dates" , SUM (time_paid::numeric)::VARCHAR "timePaid"
   FROM jobs WHERE "user_id" = $1
   GROUP BY "dates"
   ORDER BY "dates" DESC
@@ -23,7 +23,7 @@ router.get('/weekly/:id', (req, res) => {
 router.get('/daily/:id', (req, res) => {
   const userId = req.user.id
   queryText = `SELECT to_Char(date_trunc('day', date ), 'MM/DD') AS "dailyhours", 
-  SUM(time_paid::INT)::VARCHAR"timePaid" 
+  SUM(time_paid::numeric)::VARCHAR"timePaid" 
   FROM jobs WHERE "user_id" = $1
   GROUP BY dailyhours
   ORDER BY dailyhours DESC 
@@ -40,7 +40,7 @@ router.get('/daily/:id', (req, res) => {
 router.get('/monthly/:id', (req, res) => {
   const userId = req.user.id;
   queryText = `SELECT to_Char(date_trunc('month', date ), 'MM/DD') AS "monthlydates", 
-  SUM(time_paid::INT)::VARCHAR "timePaid"
+  SUM(time_paid::numeric)::VARCHAR "timePaid"
   FROM jobs WHERE "user_id" = $1
   GROUP BY monthlydates
   ORDER BY monthlydates DESC
